backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pillow==11.0.0
pytesseract==0.3.13
pydantic==2.9.2
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.26.4

# HIGHLY RECOMMENDED: For much better OCR results
openai==1.54.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
